requirements.txt
biopython
biosimulators_cobrapy
biosimulators_utils[escher] >= 0.1.150
cement
pyyaml
python_dateutil
requests_cache
setuptools
